+++ packages/suite/package.json
@@ -9,7 +9,7 @@
         "translations:extract": "rimraf translations/extractedMessages && cross-env NODE_ENV=translations babel src/**/*.{tsx,ts} > /dev/null && node ../translations-manager/lib/cli.js merge-msgs && node ../translations-manager/lib/cli.js build-csv",
         "translations:upload": "yarn translations:extract && node ../translations-manager/lib/cli.js upload",
         "translations:download": "yarn ttm build-translations && sleep 1 && yarn ttm export-translations",
-        "translations:unused": "ts-node -T -O '{\"module\": \"commonjs\"}' ../../ci/scripts/find-unused-messages.ts",
+        "translations:unused": "yarn ttm find-unused-messages --messages ./translations/master.json --src ./src",
         "type-check": "tsc --project tsconfig.json",
         "test": "run-s test:*",
         "test:unit": "jest",
