+++ packages/translations-manager/index.d.ts
@@ -1,0 +1,4 @@
+declare module 'recursive-readdir-sync' {
+    function fn(path: string): string[];
+    export = fn;
+}
