+++ packages/translations-manager/tsconfig.json
@@ -15,7 +15,7 @@
         "strictNullChecks": true,
         "outDir": "./lib"
     },
-    "include": ["src/**/*"],
+    "include": ["src/**/*", "index.d.ts"],
     "exclude": [
         "node_modules",
         "**/node_modules",
